Rename payment components files
