+++ projects/magento/pageObjects/checkout/AdyenGivingComponentsMagento.js
