+++ projects/magento/pageObjects/checkout/BancontactCardComponentsMagento.js
