+++ projects/magento/pageObjects/checkout/BoletoComponentsMagento.js
