+++ projects/magento/pageObjects/checkout/ClearPayComponentsMagento.js
