+++ projects/magento/pageObjects/checkout/GenericGiftCardComponentsMagento.js
