+++ projects/magento/pageObjects/checkout/KlarnaPayLaterComponentsMagento.js
