+++ projects/magento/pageObjects/checkout/KlarnaPayNowComponentsMagento.js
