+++ projects/magento/pageObjects/checkout/MultiBancoComponentsMagento.js
