+++ projects/magento/pageObjects/checkout/OneyComponentsMagento.js
