+++ projects/magento/pageObjects/checkout/PayPalComponentsMagento.js
